DEP-3842: Created EHL DAv3 page
