--- app/controllers/CuiController/CuiController.scala
@@ -1,36 +1,0 @@
-/*
- * Copyright 2022 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package controllers.CuiController
-
-import config.AppConfig
-import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
-import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
-import views.html.CUIViews.EmployerEnquiriesCUIView
-import javax.inject.{Inject, Singleton}
-import scala.concurrent.Future
-
-@Singleton
-class CuiController @Inject()(appConfig: AppConfig,
-                              mcc: MessagesControllerComponents,
-                              employerEnquiriesCUIView: EmployerEnquiriesCUIView) extends FrontendController(mcc) {
-
-  implicit val config: AppConfig = appConfig
-
-  def employerEnquiries: Action[AnyContent] = Action.async { implicit request =>
-      Future.successful(Ok(employerEnquiriesCUIView()))
-  }
-}
