+++ app/views/CIAPIViews/EmployerEnquiriesCUIView.scala.html
@@ -1,0 +1,39 @@
+@*
+ * Copyright 2022 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *@
+
+@import config.AppConfig
+@import templates.MainTemplateCUIv3
+@import views.html.helper.CSPNonce
+
+@this(main_template: MainTemplateCUIv3, appConfig: AppConfig)
+
+@()(implicit request: Request[_], messages: play.api.i18n.Messages)
+
+@main_template(title = messages("cui.employer.enquiries.title")) {
+    <div class="cui-wrapper">
+        <div id="cui-messaging-container">
+            <div id="nuanMessagingFrame">
+                <div id="error-message" aria-live="assertive" role="alert"></div>
+            </div>
+            <div id="cui-loading-animation">
+                <img id="cui-loading-animation-image" src='@routes.Assets.versioned("media/loading.gif")' alt="Chat loading...">
+                <div id="cui-loading-animation-text">
+                    <p>Chat loading...</p>
+                </div>
+            </div>
+        </div>
+    </div>
+}
