+++ conf/app.routes
@@ -36,8 +36,6 @@ GET         /webchat/payment-problems                                           
 
 GET         /accessibility-statement                                                    controllers.AccessibilityStatementController.accessibility(userAction: Option[String])
 
-GET         /chat/employer-enquiries                                                    controllers.CuiController.CuiController.employerEnquiries
-
 GET         /chat/ask-hmrc-online                                                       controllers.CiapiController.CiapiController.askHmrcOnline
 GET         /chat/customs-and-international-trade                                       controllers.CiapiController.CiapiController.customsInternationalTrade
 GET			/chat/vat-online															controllers.CiapiController.CiapiController.vatOnline
@@ -46,3 +44,4 @@ GET         /chat/child-benefit                                                 
 GET         /chat/construction-industry-scheme                                          controllers.CiapiController.CiapiController.constructionIndustryScheme
 GET         /chat/self-assessment                                                       controllers.CiapiController.CiapiController.selfAssessment
 GET         /chat/online-services-helpdesk                                              controllers.CiapiController.CiapiController.onlineServicesHelpdesk
+GET         /chat/employer-enquiries                                                    controllers.CiapiController.CiapiController.employerEnquiries
