+++ conf/application.conf
@@ -162,4 +162,5 @@ features {
   showCISCUI = true
   showSACUI = true
   showOSHCUI = true
+  showEHLCUI = true
 }
